+++ lib/requirejs/rails/config.rb
@@ -108,13 +108,17 @@ module Requirejs::Rails
           mod['name'] = 'almond'
           mod['include'] = name
         end
+        self.rewrite_urls_in_paths self[:build_config]
       end
       self[:build_config]
     end
 
     def run_config
-      run_config = { "baseUrl" => "/assets" }
-      run_config.merge!(self.user_config).slice(*self.run_config_whitelist)
+      unless self.has_key?(:run_config)
+        self[:run_config] = { "baseUrl" => "/assets" }
+        self[:run_config].merge!(self.user_config).slice!(*self.run_config_whitelist)
+      end
+      self[:run_config]
     end
 
     def user_config=(cfg)
@@ -145,6 +149,14 @@ module Requirejs::Rails
       self.logical_asset_filter.reduce(false) do |accum, matcher|
         accum || (matcher =~ asset)
       end ? true : false
+    end
+
+    def rewrite_urls_in_paths(cfg)
+      if cfg.has_key? 'paths'
+        cfg['paths'] = cfg['paths'].each_with_object({}) do |(k, v), h|
+          h[k] = (v =~ /^https?:/) ? 'empty:' : v
+        end
+      end
     end
   end
 end